+++ apps/web/app/api/customer-card/route.ts
@@ -7,6 +7,7 @@ import { z } from "zod";
 
 import dayjs from "@calcom/dayjs";
 import { UserRepository } from "@calcom/lib/server/repository/user";
+import { timeZoneSchema } from "@calcom/lib/timeZoneSchema";
 import { userMetadata } from "@calcom/prisma/zod-utils";
 
 import { CardComponent } from "@lib/plain/card-components";
@@ -443,7 +444,7 @@ const inputSchema = z.object({
     name: z.string().optional(),
     email: z.string().email(),
     username: z.string().optional(),
-    timeZone: z.string().optional(),
+    timeZone: timeZoneSchema.optional(),
     emailVerified: z.boolean().optional(),
     identityProvider: z.string().optional(),
     twoFactorEnabled: z.boolean().optional(),
