+++ packages/features/ee/cal-ai-phone/zod-utils.ts
@@ -1,6 +1,8 @@
 import { isValidPhoneNumber } from "libphonenumber-js";
 import z from "zod";
 
+import { timeZoneSchema } from "@calcom/lib/timeZoneSchema";
+
 export const templateTypeEnum = z.enum(["CHECK_IN_APPOINTMENT", "CUSTOM_TEMPLATE"]);
 
 const requiredFields = z.object({
@@ -143,7 +145,7 @@ export const ZGetRetellLLMSchema = z
           type: z.string(),
           cal_api_key: z.string().optional(),
           event_type_id: z.number().optional(),
-          timezone: z.string().optional(),
+          timezone: timeZoneSchema.optional(),
         })
         .passthrough()
     ),
@@ -159,7 +161,7 @@ export const ZGetRetellLLMSchema = z
                   type: z.string(),
                   cal_api_key: z.string().optional(),
                   event_type_id: z.number().optional(),
-                  timezone: z.string().optional(),
+                  timezone: timeZoneSchema.optional(),
                 })
                 .passthrough()
             ),
