+++ packages/features/users/components/UserTable/EditSheet/EditUserForm.tsx
@@ -8,6 +8,7 @@ import { z } from "zod";
 import { TimezoneSelect } from "@calcom/features/components/timezone-select";
 import { emailSchema } from "@calcom/lib/emailSchema";
 import { useLocale } from "@calcom/lib/hooks/useLocale";
+import { timeZoneSchema } from "@calcom/lib/timeZoneSchema";
 import { MembershipRole } from "@calcom/prisma/enums";
 import { trpc, type RouterOutputs } from "@calcom/trpc/react";
 import { Avatar } from "@calcom/ui/components/avatar";
@@ -58,7 +59,7 @@ const editSchema = z.object({
   avatar: z.string(),
   bio: z.string(),
   role: z.enum([MembershipRole.MEMBER, MembershipRole.ADMIN, MembershipRole.OWNER]),
-  timeZone: z.string(),
+  timeZone: timeZoneSchema,
   // schedules: z.array(z.string()),
   // teams: z.array(z.string()),
   attributes: z.array(attributeSchema).optional(),
