+++ packages/lib/date-fns/index.ts
@@ -31,22 +31,6 @@ export const formatTime = (
 };
 
 /**
- * Checks if a provided timeZone string is recognized as a valid timezone by dayjs.
- *
- * @param {string} timeZone - The timezone string to be verified.
- * @returns {boolean} - Returns 'true' if the provided timezone string is recognized as a valid timezone by dayjs. Otherwise, returns 'false'.
- *
- */
-export const isSupportedTimeZone = (timeZone: string) => {
-  try {
-    dayjs().tz(timeZone);
-    return true;
-  } catch (error) {
-    return false;
-  }
-};
-
-/**
  * Returns a localized and translated date or time, based on the native
  * Intl.DateTimeFormat available to JS. Undefined values mean the browser's
  * locale will be used.
