+++ packages/lib/date-fns/timeZoneSchema.ts
@@ -1,0 +1,8 @@
+import { z } from "zod";
+
+import { isSupportedTimeZone } from "./isSupportedTimeZone";
+
+// Schema for validating IANA timezone strings compatible with Intl.DateTimeFormat
+export const timeZoneSchema = z.string().refine((timeZone) => isSupportedTimeZone(timeZone), {
+  message: "Invalid timezone. Must be a valid IANA timezone string.",
+});
