+++ packages/platform/types/me.ts
@@ -1,12 +1,14 @@
 import { z } from "zod";
 
+import { timeZoneSchema } from "@calcom/lib/timeZoneSchema";
+
 export const userSchemaResponse = z.object({
   id: z.number().int(),
   email: z.string(),
   timeFormat: z.number().int().default(12),
   defaultScheduleId: z.number().int().nullable(),
   weekStart: z.string(),
-  timeZone: z.string().default("Europe/London"),
+  timeZone: timeZoneSchema.default("Europe/London"),
   username: z.string(),
   organizationId: z.number().nullable(),
   organization: z.object({ isPlatform: z.boolean(), id: z.number() }).optional(),
