+++ packages/platform/types/schedules/schedules-2024-04-15/schedules.ts
@@ -4,11 +4,13 @@ import { IsString, IsBoolean, IsOptional, ValidateNested, IsArray, IsDate } from
 import { DateTime } from "luxon";
 import { z } from "zod";
 
+import { timeZoneSchema } from "@calcom/lib/timeZoneSchema";
+
 const scheduleSchema_2024_04_15 = z.object({
   id: z.number().int(),
   userId: z.number().int(),
   name: z.string(),
-  timeZone: z.string().nullish(),
+  timeZone: timeZoneSchema.nullish(),
 });
 
 const availabilitySchema_2024_04_15 = z.object({
