+++ packages/trpc/server/routers/viewer/teams/getMemberAvailability.schema.ts
@@ -1,9 +1,11 @@
 import { z } from "zod";
 
+import { timeZoneSchema } from "@calcom/lib/timeZoneSchema";
+
 export const ZGetMemberAvailabilityInputSchema = z.object({
   teamId: z.number(),
   memberId: z.number(),
-  timezone: z.string(),
+  timezone: timeZoneSchema,
   dateFrom: z.string(),
   dateTo: z.string(),
 });
